requirements.txt
aiogram==2.25.1
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
chess==1.10.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
currency.converter==0.5.5
CurrencyConverter==0.17.9
cycler==0.11.0
disnake==2.9.0
dnspython==2.4.1
filelock==3.12.2
fonttools==4.42.0
frozenlist==1.4.0
glfw==2.6.2
idna==3.4
imageio==2.31.1
imgui==2.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
lazy_loader==0.3
libretranslatepy==2.1.1
lxml==4.9.3
magic-filter==1.0.11
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
nextcord==2.5.0
ninja==1.11.1
numpy==1.25.2
openai==0.27.8
opencv-python-headless==4.8.0.76
packaging==23.1
Pillow==10.0.0
pip==23.1.2
pyclipper==1.3.0.post4
pymongo==4.4.1
PyOpenGL==3.1.7
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyTelegramBotAPI==4.12.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
setuptools==65.5.0
shapely==2.0.1
six==1.16.0
sympy==1.12
telebot==0.0.5
tifffile==2023.8.12
timeago==1.0.16
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
translate==3.6.1
typing_extensions==4.7.1
urllib3==2.0.4
vacefron.py==2.0.3
yarl==1.9.2